+++ NEWS.md
@@ -49,3 +49,5 @@ SHELF v1.2.3.9001 (2017-04-13)
 * Bug fixed: ensure solid line used for linear pool when plotting. Option in plotfit added to plot all individual densities with same colour, to simplify legend.
 
 * New function linearPoolDensity for extracting density values from the linear pool.
+
+* Bug fixed: can now accept more than 26 experts.